+++ tests/testthat/test-apportion.R
@@ -33,6 +33,11 @@ test_that("apportion.R does apportion cases that should be apportioned", {
   expect_equal(apportion("mssa", NA_character_, NA_character_,
                          date_admitted = lubridate::dmy("01-01-2013"), specimen_date =
                            lubridate::dmy("05-01-2013"), date_entered = lubridate::dmy("26-10-2013")), 1)
+
+  # just for coverage stats
+  expect_equal(apportion("cdi", NA_character_, NA_character_,
+                         date_admitted = lubridate::dmy("01-01-2013"), specimen_date =
+                           lubridate::dmy("05-01-2013"), date_entered = lubridate::dmy("26-10-2013")), 1)
 })
 
 test_that("apportion.R does not trust apportion records less than two days after admission", {